requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
